Adding User Ptr support in libdrm to create/wrap a Gem object around a User Space Ptr.
BZ: 103639

A new ioctl 'DRM_IOCTL_GEM_VMAP' has been added through which a new GEM object can be created
around a memory already allocated by the User/App. This way a direct buffer sharing can be
done with the Graphics hardware without doing a copy to a memory allocated by Graphics Driver.
For example the buffer from a Camera App, containing the captured Raw data of a video frame or
a still image, can be directly used by the Gfx driver for the encode.
As App can deallocate the memory any time, the buffer destruction/synchronization is also handled.

Change-Id: Ie5d2e8c2551a9f7cab2786bf51e36f6957748209
Signed-off-by: akashgoe <akash.goel@intel.com>
Reviewed-on: http://android.intel.com:8080/107740
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Satyanantha, Rama Gopal M <rama.gopal.m.satyanantha@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>