[Graphics]: potantial fix for tombstone crash in Libdrm
BZ: 139209

By decode the test result of MTBF, we found some drm_intel_bo is already freed,
but it will be unreference again. And when called function drm_intel_gem_bo_unreference_locked_timed,
the bo_gem is already NULL, but it still use &bo_gem->refcount.
So it will crash and happend tombstone.
This is a potantial fix for crash in Libdrm.

Change-Id: Ia1dfca2883b3e92dcfd49c06151384ce16595ded
Signed-off-by: Shi Yang <yang.a.shi@intel.com>
Reviewed-on: http://android.intel.com:8080/134488
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>