external/drm: Added a flag to wait for the last write or the last read
BZ: 221178

The GEM_WAIT ioctl now accepts a flag for choosing whether to wait for
completion of the the last read or the last write performed on the
buffer.

This is part of a fix for a performance regression recently reported
on the Quadrant 2D benchmark. Apply together with drm/i915 kernel
changes and the HWC change for using the new flag.

Change-Id: Ib6a64d58de7c54d39eecdcabebaeab074841815e
Signed-off-by: Abel Bernabeu <abel.bernabeu@intel.com>
Reviewed-on: https://android.intel.com:443/235770